[fix] 자료형 long에서 double로 수정

See merge request s14-fintech-finance-sub1/S14P21D102!175

+++ frontend/WYE/app/src/main/java/com/d102/wye/data/mapper/EtfMapper.kt
@@ -32,10 +32,10 @@ fun EtfListItemResponse.toDomain() = Etf(
 fun EtfDetailResponse.toDomain() = EtfDetail(
     ticker = ticker,
     name = name,
-    currentPrice = currentPrice,
-    dailyFluctuation = dailyFluctuation,
+    currentPrice = currentPrice.toLong(),
+    dailyFluctuation = dailyFluctuation.toLong(),
     dailyFluctuationRatio = dailyFluctuationRatio,
-    volume = volume,
+    volume = volume.toLong(),
     company = company,
     riskGrade = riskGrade,
     riskType = riskType,
@@ -43,10 +43,10 @@ fun EtfDetailResponse.toDomain() = EtfDetail(
     per = per,
     pbr = pbr,
     roe = roe,
-    aum = aum,
+    aum = aum.toLong(),
     listingDate = listingDate,
-    inav = inav,
-    inavChangeAmount = inavChangeAmount,
+    inav = inav.toLong(),
+    inavChangeAmount = inavChangeAmount.toLong(),
     inavChangeRate = inavChangeRate,
 )
